% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{bitaliasing}
\alias{bitaliasing}
\title{Bitaliasing of a CRP matrix}
\usage{
bitaliasing(crps)
}
\arguments{
\item{crps}{A 2D CRP matrix}
}
\description{
Unlike \code{Uniformity}, it does not make sense to calculate the bitaliasing of a single bit vector. Thus, \code{bitaliasing} only works on 2D matrix.
}
\details{
Like \code{Unformity}, this function assumes that the CRPs are supplied in a 2D matrix, where each row corresponds to a device and each column corresponds to a CRP.

Bitaliasing measures the distribution of 1s and 0s for a single CRPs across all devices.

\deqn{Bitaliasing = \frac{1}{\#D} \sum_{d = 0}^{\#D} crp_d}
}
\examples{
## Bitaliasing of a matrix
bitaliasing(rbits(c(5, 50)))
}
\seealso{
\link[=hamming_weight]{hamming_weight}
}
