% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compare_with_ref}
\alias{compare_with_ref}
\title{Compare all values in a vector with the reference index}
\usage{
compare_with_ref(x, ref = 1)
}
\arguments{
\item{x}{Bit vector to check}

\item{ref}{The index of the reference. By default 1}
}
\description{
If a value is equal to the refence, 1 is returned and 0 otherwise.
}
\details{
TODO: Maybe return the list of comparison to calculate the mean and sd
}
