% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crps_to_df}
\alias{crps_to_df}
\title{Convert a 2D CRP matrix or 3D CRP array to a dataframe}
\usage{
crps_to_df(crps)
}
\arguments{
\item{crps}{A 2D CRP matrix or 3D CRP array}
}
\value{
A dataframe with the columns \code{device}, \code{challenge} and \code{response} if \code{crps} is a 2D matrix. If \code{crps} is a 3D array, an additional column \code{sample} is added corresponding to the 3rd dimension.
}
\description{
Convert a 2D CRP matrix or 3D CRP array to a dataframe
}
\examples{
## CRP matrix
crps <- rbits(c(2, 5))
crps_to_df(crps)

## CRP array
crps <- rbits(c(2, 5, 3))
crps_to_df(crps)
}
