% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{crps_weight}
\alias{crps_weight}
\title{Hamming weight of CRPs}
\usage{
crps_weight(crps, axis = 1)
}
\arguments{
\item{crps}{A binary vector or 2D matrix.}

\item{axis}{The axis to calculate the Hamming weight. 1 for rows and 2 for columns.}

\item{weight_fn}{Function used to calculate the weight of the CRPs}
}
\value{
The Hamming weight of the CRPs
}
\description{
This function is a wrapper around \link[=hamming_weight]{hamming_weight} to be used on vectors and 2D matrix.

This function assumes that if the CRPs are supplied in a 2D matrix, each row corresponds to a device and each column corresponds to a CRP. If another convetion is used, modify the parameter \code{axis} accordingly.
}
\details{
\subsection{Uniformity}{

Uniformity measures the distribution of 1s and 0s across all CRPs for each device. To calculate uniformity, \code{axis} should be 1.

\deqn{Uniformity = \frac{1}{\#C} \sum_{c = 0}^{\#C} crp_c}
}

\subsection{Bitaliasing}{

Bitaliasing measures the distribution of 1s and 0s for a single CRPs across all devices. To calculate bitaliasing, \code{axis} should be 2.

\deqn{Bitaliasing = \frac{1}{\#D} \sum_{d = 0}^{\#D} crp_d}
}
}
\examples{
## Hamming weight of a binary vector
v <- rbits(50)
crps_weight(v)

## Uniformity of a set of CRPs
mat <- matrix(rbits(500), nrow = 5, ncol = 10)
crps_weight(mat, 1)

## Bitaliasing of a set of CRPs
crps_weight(mat, 2)
}
\seealso{
\link[=hamming_weight]{hamming_weight}
}
