% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_crps}
\alias{df_to_crps}
\title{Convert a data frame to a CRP matrix}
\usage{
df_to_crps(df)
}
\arguments{
\item{df}{The dataframe with columns \code{device}, \code{challenge}, \code{response} and optional \code{sample}}
}
\value{
If the column \code{sample} is in the data frame, a 3D array. If not, a 2D matrix.
}
\description{
The dataframe needs to contain the columns \code{device}, \code{challenge} and \code{response}. The column \code{sample} is optional and if found, will be used to create a 3D array. Other additional columns are ignored.
}
\examples{
## Dataframe without samples
df <- expand.grid(device = 1:2, challenge = 1:5)
df$response <- rbits(2 * 5)
df_to_crps(df)

## Dataframe with samples
df <- expand.grid(device = 1:2, challenge = 1:5, sample = 1:3)
df$response <- rbits(2 * 5 * 3)
df_to_crps(df)
}
