% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{entropy_bits}
\alias{entropy_bits}
\title{Shannon entropy for binary vectors}
\usage{
entropy_bits(v)
}
\arguments{
\item{v}{A binary vector}
}
\value{
The Shannon entropy of the vector
}
\description{
Shannon entropy for binary vectors
}
\details{
The probability \eqn{p(1)} corresponds to the normalized \link{hamming_weight} of the vector.
}
\examples{
entropy_bits(c(0, 0, 0))
entropy_bits(c(1, 1, 1))
entropy_bits(rbits(20))
}
\seealso{
\link[=entropy_shannon]{entropy_shannon}
}
