% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{entropy_lgl}
\alias{entropy_lgl}
\title{Optimized Shannon entropy for logical vectors}
\usage{
entropy_lgl(v)
}
\arguments{
\item{v}{A logical or numeric vector}
}
\value{
The Shannon entropy of the vector
}
\description{
Optimized Shannon entropy for logical vectors
}
\examples{
entropy_lgl(c(0, 0, 0))
entropy_lgl(sample(c(0, 1), 20, replace = TRUE))
}
