% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entropy_p}
\alias{entropy_p}
\title{Shannon entropy for probabilities}
\usage{
entropy_p(v)
}
\arguments{
\item{v}{A vector of probabilities}
}
\value{
The entropy of the vector
}
\description{
Shannon entropy for probabilities
}
\seealso{
\link[=entropy_shannon]{entropy_shannon}
}
