% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entropy_shannon}
\alias{entropy_shannon}
\title{Shannon entropy for numeric values}
\usage{
entropy_shannon(v)
}
\arguments{
\item{v}{A vector of values}
}
\value{
The Shannon entropy of the vector
}
\description{
The Shannon entropy of a vector is calculated as:
\deqn{H(X) := -\sum_{x \in \chi} p(x) \log p(x) = E[-\log p(x)]}
}
\details{
When calculating the probabilities, it is assumed by convention that \eqn{(0\cdot \log 0 = 0)} and \eqn{(1 \cdot \log 1 = 1)}.
}
