% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{equal_to_idx}
\alias{equal_to_idx}
\title{Compare all values in a vector to the reference index}
\usage{
equal_to_idx(x, ref = 1)
}
\arguments{
\item{x}{A numeric or binary vector}

\item{ref}{The index of the value used as reference. By default \code{1}.}
}
\value{
A numeric vector containing 1 in the positions with equal values to the reference and 0 otherwise.
}
\description{
Compare all values in a vector to the reference index
}
\examples{
v <- rbits(10)
v
equal_to_idx(v, 1)
}
