% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hamming_weight}
\alias{hamming_weight}
\title{Hamming weight of a vector}
\usage{
hamming_weight(v)
}
\arguments{
\item{v}{A logical or numeric vector}
}
\value{
The hamming weight
}
\description{
The hamming weight is the number of non null symbols in a vector. In this case, the number of 1s.
}
\examples{
hamming_weight(c(0, 1, 1))
hamming_weight(c(TRUE, TRUE, TRUE))
}
