% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{intra_hd}
\alias{intra_hd}
\title{Intra Hamming distance of CRPs}
\usage{
intra_hd(crps, ref_sample = 1)
}
\arguments{
\item{crps}{A binary 2D matrix or 3D array}

\item{ref_sample}{Row of the 2D matrix used as reference CRPs}
}
\value{
The average intra Hamming distance
}
\description{
The intra Hamming distance of two sets of CRPs measures the numbers of CRPs that differ. N sets of CRPs are equal if their intra Hamming distances are 0.

A response is reliable if it does not change in time, thus, its intra Hamming distance is equal to 0.
}
\examples{
## Set of CRPs
mat <- matrix(0, nrow = 10, ncol = 50)
intra_hd(mat, 1)

## Set of devices with their respective samples
mat <- array(rbits(10 * 50 * 3), dim = c(10, 50, 3))
colMeans(intra_hd(mat))
}
\seealso{
\link[=hamming_dist]{hamming_dist}
}
