% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{length.pufmetrics}
\alias{length.pufmetrics}
\title{Returns the length of the metrics}
\usage{
\method{length}{pufmetrics}(x)
}
\arguments{
\item{x}{The PUF metrics}
}
\value{
The total number of responses
}
\description{
The length of the metrics corresponds to the total number of responses
}
\examples{
crps <- rbits(c(5, 10))
length(metrics(crps))

crps <- rbits(c(5, 10, 3))
length(metrics(crps))
}
