% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Compute the quality metrics of the given CRP matrix}
\usage{
metrics(crps, with_entropy = FALSE, ...)
}
\arguments{
\item{crps}{A 2D or 3D CRP matrix}

\item{with_entropy}{If \code{TRUE}, wrap the results of the metrics with the Shannon entropy. By default \code{FALSE}.}
}
\value{
The calculated metrics in a \code{pufmetrics} object.
}
\description{
Compute the quality metrics of the given CRP matrix
}
