% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_apply}
\alias{par_apply}
\title{Automatic parallelization of apply}
\usage{
par_apply(v, margin, fn, ...)
}
\arguments{
\item{v}{Vector of values}

\item{margin}{Vector giving the subscripts which the function will be applied over. E.g., for a matrix 1 indicates rows, 2 indicates columns, c(1, 2) indicates rows and columns. Where X has named dimnames, it can be a character vector selecting dimension names.}

\item{fn}{Function to apply.}

\item{...}{Rest of arguments passed to \code{fn}.}
}
\value{
Vector or array or list of values obtained by applying a function to margins of an array or matrix.
}
\description{
Use \link[parallel:clusterApply]{parallel::parApply} if a parallel context has been created with \link{register_parallel}. Otherwise use \link[base:apply]{base::apply}.
}
\seealso{
\link{register_parallel}, \link[parallel:clusterApply]{parallel::parApply}, \link{apply}
}
