% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_vapply}
\alias{par_vapply}
\title{Automatic parallelization of vapply}
\usage{
par_vapply(X, FUN, FUN.VALUE = NULL, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{FUN.VALUE}{a (generalized) vector; a template for the return
    value from FUN.  See \sQuote{Details}.}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
#' Use \link[parallel:clusterApply]{parallel::parApply} if a parallel context has been created with \link{register_parallel}. Otherwise use \link[base:apply]{base::apply}.
}
\seealso{
\link{register_parallel}, \link[parallel:clusterApply]{parallel::parSapply}
}
