% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ratio_bits}
\alias{ratio_bits}
\title{Ratio of bits in a vector}
\usage{
ratio_bits(v, abs = TRUE)
}
\arguments{
\item{v}{A logical or numeric vector}

\item{abs}{If `TRUE` (default) the ratio is absolute}
}
\value{
The ratio of bits
}
\description{
The ratio is calculated as the number of ones minus the number of zeros.
}
\details{
\deqn{Ratio = \frac{\sum_{i=0}^{n} v_i}{n}}

If the ratio is not absolute, a positive result indicates that there are more 1s than 0s, while a negative results indicates the opposite.
}
\examples{
ratio_bits(c(TRUE, FALSE, TRUE), abs = TRUE)
ratio_bits(c(0, 1, 0), abs = FALSE)
ratio_bits(c(0, 1, 0, 1))
}
