% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbits}
\alias{rbits}
\title{Random binary vector}
\usage{
rbits(size, p = 0.5)
}
\arguments{
\item{size}{The size of the vector}

\item{p}{Probability of obtaining a 1. By default is \code{0.5}.}
}
\value{
The generated binary vector
}
\description{
This function is a wrapper around \link{sample} to generate binary vectors.
}
\examples{
## Unbiased probabilities
rbits(10)

## Biased probabilities
rbits(10, p = 0.8)
}
