% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{register_parallel}
\alias{register_parallel}
\title{Register a parallel cluster}
\usage{
register_parallel(ctx)
}
\arguments{
\item{ctx}{A cluster created with \code{\link[parallel:makeCluster]{parallel::makeCluster()}}}
}
\description{
The user should create the cluster and close it.
}
\examples{
\dontrun{
ctx <- parallel::makeCluster(2)
register_parallel(ctx)
}
}
\seealso{
\code{\link[parallel:makeCluster]{parallel::makeCluster()}}
}
