% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{reliability}
\alias{reliability}
\title{Reliability of CRPs}
\usage{
reliability(crps, ref = 1)
}
\arguments{
\item{crps}{A binary vector, 2D matrix or 3D array.}

\item{ref}{Numeric index for the reference sample: If \code{crps} is a vector, is the index of the reference sample; If \code{crps} is a 2D matrix, the row to use as reference; If \code{crps} is a 3D array,the row for all 3rd dimension matrix.}
}
\description{
The reliability of a PUF is defined as \eqn{1 - Intra_{HD}}
}
\examples{
## Set of CRPs
mat <- matrix(rbits(200), nrow = 10, ncol = 20)
intra <- intra_hd(mat, 1)
all(1 - intra == reliability(mat, 1))
}
\seealso{
\link[=intra_hd]{intra_hd}
}
