% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{summary.pufmetrics}
\alias{summary.pufmetrics}
\title{Create a summary of the metrics}
\usage{
\method{summary}{pufmetrics}(object, ...)
}
\arguments{
\item{object}{The PUF metrics}

\item{...}{Rest of the parameters}
}
\value{
The total number of responses
}
\description{
Create a summary of the metrics
}
\examples{
arr <- rbits(c(5, 10, 3))
summary(metrics(arr))
}
