% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uniqueness}
\alias{uniqueness}
\title{Uniqueness of a CRP table}
\usage{
uniqueness(crps)
}
\arguments{
\item{crps}{A logical or numeric matrix}
}
\value{
The uniqueness of the CRP table
}
\description{
The uniqueness is calculated as the average of the hamming distance of the CRPs of two devices, for every pair of devices.
}
\examples{
mat <- matrix(sample(c(0, 1, 100, replace = TRUE)), nrow = 10, ncol = 10)
uniqueness(mat)
}
