% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{with_entropy}
\alias{with_entropy}
\title{Convert metrics to entropy calculation}
\usage{
with_entropy(metrics)
}
\arguments{
\item{metrics}{A pufmetrics object}
}
\value{
The new pufmetrics object where the metrics have been passed to \code{entropy_p} description
}
\description{
Convert metrics to entropy calculation
}
\examples{
crps <- rbits(c(5, 50))
m <- metrics(crps)
m
with_entropy(m)
}
\seealso{
\link[=entropy_p]{entropy_p}
}
