% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_pairwise}
\alias{compare_pairwise}
\title{Compare a matrix by pairs of rows}
\usage{
compare_pairwise(m, fn, ...)
}
\arguments{
\item{m}{Vector of values.}

\item{fn}{Function that receives two rows.}

\item{...}{Rest of arguments passed to \code{fn}.}
}
\value{
Vector containing the results of all comparisons.
}
\description{
Each pair of rows is compared using the given function.
The pairs are chosen without repetition.
}
\examples{
#' Compare a matrix by pairs of rows
m <- matrix(rbits(25), 5, 5)
res <- compare_pairwise(m, hamming_dist, norm = TRUE)
res
length(res) == (5 * 4 / 2)

## Equivalence to uniqueness
res <- compare_pairwise(m, function(f, s) 1 - hamming_dist(f, s, norm = TRUE))
all(uniqueness(m) == res)
}
