% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{entropy_p}
\alias{entropy_p}
\title{Shannon entropy for probabilities}
\usage{
entropy_p(v)
}
\arguments{
\item{v}{A vector of probabilities where each probability is treated as P(1)}
}
\value{
The Shannon entropy of each probability
}
\description{
Shannon entropy for probabilities
}
\seealso{
\link[=entropy_shannon]{entropy_shannon}
}
