% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hamming_dist}
\alias{hamming_dist}
\alias{\%<>\%}
\alias{\%</>\%}
\title{Hamming distance of two vectors}
\usage{
hamming_dist(x, y, norm = FALSE)

x \%<>\% y

x \%</>\% y
}
\arguments{
\item{x}{A numeric or logical vector}

\item{y}{A numeric or logical vector}

\item{norm}{If \code{TRUE} (default is \code{FALSE}) normalize the distance to the vector length}
}
\value{
The hamming distance
}
\description{
Hamming distance of two vectors
}
\examples{
hamming_dist(c(0, 1, 0), c(0, 0, 0))
hamming_dist(c(0, 1, 0), c(0, 0, 0), norm = TRUE)
c(0, 1, 0) \%<>\% c(1, 0, 0)
c(0, 1, 0) \%</>\% c(1, 0, 0)
}
