% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hamming_dist}
\alias{hamming_dist}
\alias{\%HD\%}
\alias{\%NHD\%}
\title{Hamming distance of two vectors}
\usage{
hamming_dist(x, y, norm = FALSE)

x \%HD\% y

x \%NHD\% y
}
\arguments{
\item{x}{A numeric or logical vector}

\item{y}{A numeric or logical vector}

\item{norm}{If \code{TRUE} normalize the distance to the vector length. By default it's \code{FALSE}}
}
\value{
The Hamming distance
}
\description{
The Hamming distance of two vectors corresponds to the number of positions where the values differ.
}
\details{
NAs are discarded in any of the vectors by using \code{na.rm = TRUE} in \code{sum}.
}
\examples{
hamming_dist(c(0, 1, 0), c(0, 0, 0))
hamming_dist(c(0, 1, 0), c(0, 0, 0), norm = TRUE)

## NAs in any of the vectors are discarded
hamming_dist(c(1, NA, 3, 4, 5), c(1, 2, NA, 4, 5))
c(0, 1, 0) \%HD\% c(1, 0, 0)
c(0, 1, 0) \%NHD\% c(1, 0, 0)
}
