% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_vapply}
\alias{par_vapply}
\title{Automatic parallelization of vapply}
\usage{
par_vapply(v, fn, value = NULL, ...)
}
\arguments{
\item{v}{Vector of values.}

\item{fn}{Function to apply.}

\item{value}{Expected type of value.}

\item{...}{Rest of arguments passed to \code{fn}.}
}
\value{
Vector or array or list of values obtained by applying a function to margins of an array or matrix.
}
\description{
Use \link[parallel:clusterApply]{parallel::parApply} if a parallel context has been created with \link{register_parallel}. Otherwise use \link[base:apply]{base::apply}.
}
\seealso{
\link{register_parallel}, \link[parallel:clusterApply]{parallel::parSapply}, \link{vapply}
}
