% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot.pufmetrics}
\alias{plot.pufmetrics}
\title{Create a plot with a summary of all the metrics}
\usage{
\method{plot}{pufmetrics}(m, ...)
}
\arguments{
\item{m}{The PUF metrics}

\item{...}{Additional parameters}
}
\description{
If the metrics come from a 2D CRP table, 3 histograms are created for uniformity, bitaliasing and uniqueness.
}
\details{
In the case of a 3D CRP table, 4 graphs are created, 3 histograms for uniformity, bitaliasing and uniqueness, combining all samples, and a matrix showing the reliability of each response.
}
