% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbits}
\alias{rbits}
\title{Random binary vector}
\usage{
rbits(size, prob = NULL)
}
\arguments{
\item{size}{The size of the vector}

\item{prob}{A numeric vector with 2 values for the probabilities of \code{0} and \code{1} respectively. By default the probabilities are \code{0.5} each.}
}
\value{
The generated binary vector
}
\description{
This function is a wrapper around \link{sample} to generate binary vectors.
}
\examples{
## Unbiased probabilities
rbits(10)

## Biased probabilities
rbits(10, prob = c(0.1, 0.9))
}
