% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/formulas.R
\name{uniqueness}
\alias{uniqueness}
\alias{inter_hd}
\title{Uniqueness of CRPs}
\usage{
uniqueness(crps)

inter_hd(crps)
}
\arguments{
\item{crps}{A logical or numeric matrix}
}
\value{
The uniqueness of the CRP table
}
\description{
The uniqueness is calculated as the average of the hamming distance of the CRPs of two devices, for every pair of devices.
The number of pairs of devices is calculated with the following formula, where \eqn{D} is the number of devices.
\deqn{N = \frac{D(D-1)}{2}}
}
\examples{
mat <- rbits(c(10, 10))
uniqueness(mat)
}
